requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile cookbook/worldbuilding/requirements.in
#
altair==5.2.0
    # via streamlit
annotated-types==0.6.0
    # via pydantic
anyio==4.2.0
    # via httpx
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
blinker==1.7.0
    # via streamlit
boto3==1.34.36
    # via phidata
botocore==1.34.36
    # via
    #   boto3
    #   phidata
    #   s3transfer
cachetools==5.3.2
    # via streamlit
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   streamlit
    #   typer
docker==7.0.0
    # via phidata
gitdb==4.0.11
    # via gitpython
gitpython==3.1.41
    # via
    #   phidata
    #   streamlit
h11==0.14.0
    # via httpcore
httpcore==1.0.2
    # via httpx
httpx==0.25.2
    # via
    #   ollama
    #   phidata
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==7.0.1
    # via streamlit
jinja2==3.1.3
    # via
    #   altair
    #   pydeck
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.21.1
    # via altair
jsonschema-specifications==2023.12.1
    # via jsonschema
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
numpy==1.26.4
    # via
    #   altair
    #   pandas
    #   pyarrow
    #   pydeck
    #   streamlit
ollama==0.1.6
    # via -r cookbook/worldbuilding/requirements.in
packaging==23.2
    # via
    #   altair
    #   docker
    #   streamlit
pandas==2.2.0
    # via
    #   altair
    #   streamlit
phidata==2.3.41
    # via -r cookbook/worldbuilding/requirements.in
pillow==10.2.0
    # via streamlit
protobuf==4.25.2
    # via streamlit
pyarrow==15.0.0
    # via streamlit
pydantic==2.6.1
    # via
    #   phidata
    #   pydantic-settings
pydantic-core==2.16.2
    # via pydantic
pydantic-settings==2.1.0
    # via phidata
pydeck==0.8.1b0
    # via streamlit
pygments==2.17.2
    # via rich
python-dateutil==2.8.2
    # via
    #   botocore
    #   pandas
    #   streamlit
python-dotenv==1.0.1
    # via
    #   phidata
    #   pydantic-settings
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via phidata
referencing==0.33.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   docker
    #   streamlit
rich==13.7.0
    # via
    #   phidata
    #   streamlit
rpds-py==0.17.1
    # via
    #   jsonschema
    #   referencing
s3transfer==0.10.0
    # via boto3
six==1.16.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.25
    # via -r cookbook/worldbuilding/requirements.in
streamlit==1.31.0
    # via -r cookbook/worldbuilding/requirements.in
tenacity==8.2.3
    # via streamlit
toml==0.10.2
    # via streamlit
tomli==2.0.1
    # via phidata
toolz==0.12.1
    # via altair
tornado==6.4
    # via streamlit
typer==0.9.0
    # via phidata
typing-extensions==4.9.0
    # via
    #   phidata
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   streamlit
    #   typer
tzdata==2023.4
    # via pandas
tzlocal==5.2
    # via streamlit
urllib3==2.0.7
    # via
    #   botocore
    #   docker
    #   requests
validators==0.22.0
    # via streamlit
zipp==3.17.0
    # via importlib-metadata